+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/interaction/dto/shopping/cart/ChangeProductQuantityRequest.java
@@ -22,5 +22,5 @@ public class ChangeProductQuantityRequest {
 
     // Новое количество товара
     @NotNull
-    Integer newQuantity;
+    Long newQuantity;
 }
